Add --flatten-dot for easy flattening
For some reason I don't really, I made flatten use '_' as the default
joiner. this is visually annoying when keys have '_' in them already,
because it is unclear what is the parent keyname vs. the sub key. So I
added a new option -f | --flatten-dot that will do the flattening, but
use '.' as the default keyname joiner.

This, combined with --all-names will give a sense of the key heirarchy
for most objects when doing some quick exploratory analysis, e.g.

join_date
user.first_name
user.last_name